audio/external: func_800C357C func_800C9D0C

+++ src/audio/external.h
@@ -145,7 +145,7 @@ void func_800C9060(u8, u32);
 void func_800C90F4(u8, uintptr_t);
 void func_800C97C4(u8);
 void func_800C9A88(u8);
-void func_800C9D0C(s32);
+void func_800C9D0C(u8);
 void func_800C9D80(Vec3f, Vec3f, u32);
 void func_800C9EF4(Vec3f, u32);
 void func_800C9F90(u8);
